Added my (unfinished) Heap Sort implementation
One of our newest projects in class that we have begun working on is
implementation of a max heap sort.  I have added what I have completed
so far.  Unfortunately, the sorting after inserting or removing an item
doesn't completely work yet.  I know where the problem is, just a matter
of thinking through and fixing the logic.